Remove vue router
