+++ package.json
@@ -34,7 +34,6 @@
     "typescript": "^5.8.3",
     "vite": "^6.2.5",
     "vitest": "^3.1.1",
-    "vue-router": "^4.5.0",
     "vue-tsc": "^2.2.8",
     "vuetify": "^3.8.0"
   }
