--- src/app/views/PageMain.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-import Intra from "@/app/views/AreaIntra.vue";
-import Settings from "@/app/views/AreaSettings.vue";
-</script>
-
-<template>
-  <Intra/>
-  <Settings class="pt-2"/>
-</template>
