+++ tests/views/App.test.ts
@@ -1,15 +1,17 @@
 import { shallowMount } from "@vue/test-utils";
 import { test } from "vitest";
-import { RouterView } from "vue-router";
 
+import Intra from "@/app/views/AreaIntra.vue";
+import Settings from "@/app/views/AreaSettings.vue";
 import App from "@/app/views/App.vue";
-import { router } from "@/app/plugins/router";
 
 test("Mounted", ({ expect }) => {
-  using wrapper = shallowMount(App, { global: { plugins: [router] } });
+  using wrapper = shallowMount(App);
 
-  expect(wrapper.findComponent(RouterView).props()).toStrictEqual({
-    name: "default",
-    route: undefined,
-  });
+  const intra = wrapper.findComponent(Intra);
+  expect(intra.classes()).toStrictEqual([]);
+  expect(intra.props()).toStrictEqual({});
+  const settings = wrapper.findComponent(Settings);
+  expect(settings.classes()).toStrictEqual(["pt-2"]);
+  expect(settings.props()).toStrictEqual({});
 });
